backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sentence-transformers==2.5.1
torch==2.7.1
numpy==1.26.4
pydantic==2.6.1
python-dotenv==1.0.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.2.0
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
pybind11>=2.11
chromadb>=0.4.22
requests>=2.31.0